# aethel-core/internal/api/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Aethel Workspace API
  version: "1.0.0"
  description: |
    REST API for Aethel Workspace — an open-source, self-hosted e-office platform
    for institutional document workflows.

    ## Authentication

    All endpoints (except those marked **Public**) require a JWT access token:

    ```
    Authorization: Bearer <access_token>
    ```

    Obtain a token via `POST /api/v1/auth/login`. Tokens expire in **15 minutes**.
    Use `POST /api/v1/auth/refresh` to renew silently using the `refresh_token` httpOnly cookie.

    ## CSRF Protection

    State-changing requests (`POST`, `PATCH`, `PUT`, `DELETE`) from browser clients must
    include the CSRF token in a header:

    ```
    X-CSRF-Token: <value of the csrf_token cookie>
    ```

    The `csrf_token` cookie is set by the server on login. Non-browser clients (CLI, mobile)
    that do not send cookies are exempt from this requirement.

    ## Rate Limiting

    - **Global:** 600 requests/min per IP (unauthenticated)
    - **Per user:** 300 requests/min per authenticated user
    - **Login endpoint:** 20 requests/min per IP (credential-stuffing protection)

    Responses exceeding the limit receive `429 Too Many Requests` with a `Retry-After` header.

    ## Error Format

    All error responses use a consistent JSON envelope:
    ```json
    { "error": "human-readable description", "field": "fieldName (validation errors only)" }
    ```

  contact:
    name: Aethel Workspace
...

  license:
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: AuthLogin
      summary: Log in and receive tokens
      description: |
        Authenticates the user with email and password.
        On success, sets two cookies and returns the access token in the response body:
        - `refresh_token` — httpOnly; JavaScript cannot read this
        - `csrf_token` — readable; JavaScript must send this in `X-CSRF-Token` on subsequent requests

        **Account lockout:** After 5 consecutive failures, the account is locked for 15 minutes.
        The same `invalid credentials` error is returned regardless of whether the email exists
        (prevents user enumeration).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            example:
              email: "marcus.webb@ministry.gov"
              password: "S3cur3P@ssw0rd!"
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              description: Sets `refresh_token` (httpOnly) and `csrf_token` (readable) cookies
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, expires_in, role]
                properties:
                  access_token: { type: string }
                  token_type: { type: string, enum: [Bearer] }
                  expires_in:
                    { type: integer, description: Seconds until expiry }
                  role: { $ref: "#/components/schemas/UserRole" }
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI4ZjNhMTJiYy0..."
                token_type: Bearer
                expires_in: 900
                role: RECEPTION
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid email or password (same response whether email exists or not)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "invalid credentials" }
        "423": { $ref: "#/components/responses/Locked" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: AuthRefresh
      summary: Exchange refresh token for a new access token
      description: |
        Uses the `refresh_token` httpOnly cookie to issue a new access token and rotate
        the refresh token. The old refresh token is invalidated (token rotation).
        No request body required — the cookie is sent automatically by the browser.
      security: []
      responses:
        "200":
          description: New access token issued
          headers:
            Set-Cookie:
              description: New `refresh_token` and `csrf_token` cookies
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [access_token, expires_in]
                properties:
                  access_token: { type: string }
                  expires_in: { type: integer }
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expires_in: 900
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: AuthLogout
      summary: Revoke the current session
      description: |
        Invalidates the current refresh token and clears both cookies.
        The access token remains valid until its natural expiry (15 min max).
      responses:
        "204":
          description: Session revoked successfully
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/password-reset/request:
    post:
      tags: [Auth]
      operationId: AuthPasswordResetRequest
      summary: Request a password reset email
      description: |
        Sends a password reset link to the provided email address.
        Always returns `202 Accepted` — does not reveal whether the email exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
            example:
              email: "marcus.webb@ministry.gov"
      responses:
        "202":
          description: Request accepted (email sent if account exists)
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      operationId: AuthPasswordResetConfirm
      summary: Apply a password reset token
      description: |
        Validates the one-time token and sets a new password.
        Tokens expire after 1 hour and are single-use.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordResetConfirmRequest" }
            example:
              token: "abc123xyz..."
              newPassword: "N3wS3cur3P@ss!"
      responses:
        "204":
          description: Password reset successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid or expired reset token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "invalid or expired token" }

  # ── Config ────────────────────────────────────────────────────────────────────
  /config:
    get:
      tags: [Config]
      operationId: ConfigGet
      summary: Get full runtime configuration
      description: |
        Returns the complete org configuration — branding, navigation, feature flags,
        and org profile — from the in-memory config cache (5-min TTL).
        This is the primary endpoint for the Nuxt SSR frontend on page load.
      responses:
        "200":
          description: Full runtime configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppConfig" }
              example:
                branding:
                  primaryColor: "#4f46e5"
                  neutralPalette: "slate"
                  fontFamily: "Inter"
                  logoUrl: null
                  wordmark: "Aethel Workspace"
                nav:
                  - label: "Reception"
                    roles: [RECEPTION, ADMIN]
                    items:
                      - label: "Dashboard"
                        icon: "i-lucide-layout-dashboard"
                        to: "/dashboard"
                        badge: null
                features:
                  greenNotingEnabled: false
                  externalSmtpEnabled: false
                  require2faForAdmin: false
                org:
                  name: "Ministry of Internal Affairs"
                  timezone: "Asia/Ho_Chi_Minh"
                  locale: "vi-VN"
                  contactEmail: "admin@ministry.gov.vn"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /config/branding:
    get:
      tags: [Config]
      operationId: ConfigGetBranding
      summary: Get branding configuration
      description: Returns only the branding sub-object from the config cache.
      responses:
        "200":
          description: Branding configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandingConfig" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /config/nav:
    get:
      tags: [Config]
      operationId: ConfigGetNav
      summary: Get navigation structure
      description: Returns the nav groups and items. Includes any runtime overrides stored in system_settings.
      responses:
        "200":
          description: Navigation structure
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NavGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /config/features:
    get:
      tags: [Config]
      operationId: ConfigGetFeatures
      summary: Get feature flags
      description: Returns the feature flags configuration object.
      responses:
        "200":
          description: Feature flags
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeatureFlags" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /admin/config/branding:
    patch:
      tags: [Config]
      operationId: AdminConfigPatchBranding
      summary: Update branding configuration
      description: |
        Partially updates the branding configuration. All fields are optional.
        Invalidates the config cache so the next read reflects the new values.
        Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchBrandingRequest" }
            example:
              primaryColor: "#0ea5e9"
              wordmark: "Ministry E-Office"
      responses:
        "200":
          description: Updated branding configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandingConfig" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/config/nav:
    patch:
      tags: [Config]
      operationId: AdminConfigPatchNav
      summary: Update navigation structure
      description: |
        Replaces the runtime nav override stored in system_settings.
        The full nav array must be provided. Invalidates the config cache.
        Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchNavRequest" }
      responses:
        "200":
          description: Updated navigation structure
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NavGroup" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/config/features:
    patch:
      tags: [Config]
      operationId: AdminConfigPatchFeatures
      summary: Update feature flags
      description: |
        Partially updates feature flags. All fields are optional booleans.
        Invalidates the config cache. Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchFeaturesRequest" }
            example:
              greenNotingEnabled: true
      responses:
        "200":
          description: Updated feature flags
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeatureFlags" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/config/org:
    patch:
      tags: [Config]
      operationId: AdminConfigPatchOrg
      summary: Update org profile
      description: |
        Partially updates the organization profile. All fields are optional.
        Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchOrgRequest" }
            example:
              name: "Ministry of Internal Affairs"
              timezone: "Asia/Ho_Chi_Minh"
      responses:
        "200":
          description: Updated org profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrgProfile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ── Dispatch ──────────────────────────────────────────────────────────────────
  /dispatches:
    get:
      tags: [Dispatch]
      operationId: DispatchListInbox
      summary: List inbound dispatch queue
      description: Returns the active inbound queue for the current user's department. Paginated.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of inbound dispatches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/DispatchSummary" }
              example:
                data:
                  - id: "8f3a12bc-1234-5678-abcd-000000000001"
                    trackingNumber: "INB-2026-001234"
                    direction: INBOUND
                    senderName: "Tax Authority"
                    priorityLevel: PRIORITY
                    statusState: UNDER_REVIEW
                    createdAt: "2026-06-01T08:00:00Z"
                total: 42
                page: 1
                pageSize: 20
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [Dispatch]
      operationId: DispatchCreate
      summary: Log a new inbound dispatch
      description: Creates a new inbound dispatch record. Triggers automatic routing rule evaluation.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDispatchRequest" }
      responses:
        "201":
          description: Dispatch created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalError" }
  /dispatches/outbound:
    get:
      tags: [Dispatch]
      operationId: DispatchListOutbound
      summary: List outbound dispatch queue
      description: Returns the outbound queue for the current organization. Paginated.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of outbound dispatches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/DispatchSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [Dispatch]
      operationId: DispatchCreateOutbound
      summary: Log a new outbound dispatch
      description: Creates a new outbound dispatch record.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDispatchRequest" }
      responses:
        "201":
          description: Outbound dispatch created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalError" }
  /my-dispatches:
    get:
      tags: [Dispatch]
      operationId: DispatchListMine
      summary: List dispatches assigned to me
      description: Returns all dispatches currently assigned to the authenticated user.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of the user's assigned dispatches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/DispatchSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /search:
    get:
      tags: [Dispatch]
      operationId: DispatchSearch
      summary: Full-text search across dispatches
      description: |
        Searches dispatch records by tracking number, sender name, subject line, or document type.
        Uses PostgreSQL full-text search (pg_trgm).
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2 }
          description: Search query string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/DispatchSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /dispatches/{id}:
    get:
      tags: [Dispatch]
      operationId: DispatchGetByID
      summary: Get dispatch detail
      description: Returns the full dispatch record including the event timeline.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Full dispatch detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispatch" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /dispatches/{id}/status:
    patch:
      tags: [Dispatch]
      operationId: DispatchUpdateStatus
      summary: Update dispatch status
      description: Transitions the dispatch to a new status state. Logs a dispatch event.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/DispatchStatus" }
            example:
              status: IN_TRANSIT
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /dispatches/{id}/assign:
    post:
      tags: [Dispatch]
      operationId: DispatchAssign
      summary: Assign dispatch to a user or department
      description: |
        Manually assigns the dispatch to a specific user or department.
        Sets `isManuallyRouted = true`, bypassing the automatic routing rules.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignDispatchRequest" }
            example:
              userId: "8f3a12bc-1234-5678-abcd-000000000042"
              isManualOverride: true
      responses:
        "200":
          description: Dispatch assigned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispatch" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /dispatches/{id}/acknowledge:
    post:
      tags: [Dispatch]
      operationId: DispatchAcknowledge
      summary: Record receipt acknowledgement
      description: |
        Records that the recipient has acknowledged receipt of the dispatch.
        Sets `acknowledgedAt` and `acknowledgedByUserId` on the dispatch record.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Acknowledgement recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispatch" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /notifications/{id}/read:
    patch:
      tags: [Notifications]
      operationId: NotificationsMarkRead
      summary: Mark a notification as read
      description: Sets the notification's read status to true.
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Notification marked as read
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token from `POST /api/v1/auth/login`.
        Claims: `sub` (user UUID), `role`, `iat`, `exp`, `jti`.
        Expiry: 15 minutes. Use the refresh endpoint to renew.
...